requirements.txt
-f https://cdn.research.pdx.edu/pypi/dist/
Django==1.7.2
Markdown==2.5.2
PyMySQL==0.6.2
coverage==3.7.1
django-arcutils==1.2.0
django-cloak==0.1.3
django-extensions==1.4.6
django-filter==0.8
model-mommy==1.2.2
pyasn1==0.1.7
ldap3==0.9.8.7
pytz==2014.9
six==1.7.3
varlet
-e git+https://github.com/PSU-OIT-ARC/django-github-hook.git#egg=github_hook
